first pass as conf as interface, wip
