+++ cmd/bosun/conf/native/parse/test_invalid/2
